backend/requirements.txt
# Core Django packages
asgiref==3.11.0
channels==4.3.2
Django==6.0.2
djangorestframework==3.16.1

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Environment variables
python-dotenv==1.0.1

# Firebase
firebase_admin==7.1.0

# CORS
django-cors-headers==4.3.1

# Static files
whitenoise==6.6.0

# Admin UI
django-jazzmin==3.0.0

# Image handling
Pillow==11.0.0

# WebSockets (Channels dependencies)
daphne==4.1.0
channels-redis==4.2.1

# Redis & Caching
redis==5.2.1
django-redis==5.4.0

# Task Queue
celery==5.4.0
